refactor: 메인 페이지 코드 분리
